+++ src/App.tsx
@@ -24,7 +24,7 @@ import PurposeIngredientProducts from "./pages/purpose/PurposeIngredientProducts
 import ProductDetailPage from "./pages/MainProductDetailPage";
 import PurposeBrandProducts from "./pages/purpose/PurposeBrandProductsPage";
 
-import LoginPage from "./pages/auth/LoginPage"; //자체 로그인 페이지
+import LoginPage from "./pages/auth/EmailLoginPage"; //자체 로그인 페이지
 
 // 레이아웃
 import RootLayout from "./layout/RootLayout";
@@ -47,7 +47,7 @@ const router = createBrowserRouter([
         element: <SignInPage />,
       },
       {
-        path: "login2", //로그인 페이지
+        path: "login/email", //로그인 페이지
         element: <LoginPage />,
       },
       {
